backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
numpy
scikit-learn
librosa
soundfile
requests
joblib
xgboost
mangum
